Bypass transform publish if transform from base_link to odom failed
In response to Issue #198: Instead of publishing an "Identity" transform if the base_link to odom transform fails, just skip this update.
While this is ultimately covering for likely bigger problem with the system (i.e. insufficient CPU power or out of spec settings), it produces a less catastrophic side effect.